+++ src/file.js
@@ -1,5 +1,5 @@
 import fs from 'fs';
-import promisify from 'util.promisify';
+import {promisify} from 'util';
 
 const readFile = promisify(fs.readFile);