+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.2.0] - 20-12-2017
+* Refactors most of the project.
+* The first argument is now split to get the directory of the string if it is a long path name.
+* Cleans up promises and async await functions.
+* Try catch now throws errors instead of just console.erroring them.
+* Removes Babel as only supporting node version 8.0.0 and up.
+
 ## [0.1.2] - 12-12-2017
 * Remvoes postinstall from package.json
 
@@ -9,4 +16,3 @@ All notable changes to this project will be documented in this file.
 
 ## [0.1.0] - 12-12-2017
 * Refactors entire project from versions previously as now this is a glboal node package.
-